dbeaver/pro#6504: revert to last stable version
