+++ plugins/org.jkiss.dbeaver.ui.ai/src/org/jkiss/dbeaver/ui/ai/internal/AIUIMessages.java
@@ -48,11 +48,9 @@ public class AIUIMessages extends NLS {
     public static String gpt_preference_page_ai_query_confirm_rule_disable_autocommit;
 
     public static String gpt_preference_page_ai_connection_test_label;
-    public static String gpt_preference_page_ai_connection_test_model_not_chosen;
+    public static String gpt_preference_page_ai_connection_test_tooltip;
     public static String gpt_preference_page_ai_connection_test_connection_success_title;
     public static String gpt_preference_page_ai_connection_test_connection_success_message;
-    public static String gpt_preference_page_ai_connection_test_connection_warning_title;
-    public static String gpt_preference_page_ai_connection_test_connection_warning_message;
     public static String gpt_preference_page_ai_connection_test_connection_error_title;
     public static String gpt_preference_page_ai_connection_test_connection_error_message;
