+++ plugins/org.jkiss.dbeaver.ui.ai/src/org/jkiss/dbeaver/ui/ai/internal/AIUIMessages.properties
@@ -24,13 +24,11 @@ gpt_preference_page_ai_query_confirm_rule_execute = Execute immediately
 gpt_preference_page_ai_query_confirm_rule_confirm = Show confirmation
 gpt_preference_page_ai_query_confirm_rule_disable_autocommit = Disable autocommit
 gpt_preference_page_ai_connection_test_label=Test Connection ...
-gpt_preference_page_ai_connection_test_model_not_chosen=Please choose a model
+gpt_preference_page_ai_connection_test_tooltip=Only settings that have been applied will be used for the connection test
 gpt_preference_page_ai_connection_test_connection_success_title=Connection Test Success
-gpt_preference_page_ai_connection_test_connection_success_message=You are connected to the AI assistant. Chosen model: {0}
-gpt_preference_page_ai_connection_test_connection_warning_title=Connection Test Warning
-gpt_preference_page_ai_connection_test_connection_warning_message=Connection succeeded, but the chosen model {0} was not found
+gpt_preference_page_ai_connection_test_connection_success_message=You are connected to the {0} AI assistant
 gpt_preference_page_ai_connection_test_connection_error_title=Connection Test Error
-gpt_preference_page_ai_connection_test_connection_error_message=Error while connecting to the AI assistant
+gpt_preference_page_ai_connection_test_connection_error_message=Error while connecting the {0} AI assistant
 
 confirm_meta_transfer_usage_title = Transfer information to AI Vendor
 confirm_meta_transfer_usage_message = To use AI features, DBeaver needs to transfer your database metadata information (table and column names) to the AI vendor API.\n\nDo you confirm it for connection ''{0}''?
